+++ project/CasbahBuild.scala
@@ -17,7 +17,7 @@ object CasbahBuild extends Build {
     organizationHomepage := Some(url("http://www.mongodb.org")),
     version      := "2.7.5",
     scalaVersion := "2.10.4",
-    crossScalaVersions := Seq("2.11.4", "2.10.4", "2.9.3")
+    crossScalaVersions := Seq("2.11.5", "2.10.4", "2.9.3")
   )
 
   val allSourceDirectories = SettingKey[Seq[Seq[File]]]("all-source-directories")